Moved pytest fixture that returns test data to conftest.py
