+++ imap_processing/tests/idex/test_idex_l0.py
@@ -1,28 +1,9 @@
 """Tests the decommutation process for IDEX CCSDS Packets."""
 
-from pathlib import Path
-
 import numpy as np
-import pytest
 import xarray as xr
 
 from imap_processing import imap_module_directory
-from imap_processing.idex.l1.idex_l1 import PacketParser
-
-
-@pytest.fixture(scope="session")
-def decom_test_data() -> xr.Dataset:
-    """Return a ``xarray`` dataset containing test data.
-
-    Returns
-    -------
-    dataset : xarray.Dataset
-        A ``xarray`` dataset containing the test data
-    """
-    test_file = Path(
-        f"{imap_module_directory}/tests/idex/imap_idex_l0_raw_20230725_v001.pkts"
-    )
-    return PacketParser(test_file, "001").data
 
 
 def test_idex_decom_length(decom_test_data: xr.Dataset):
